# api.yaml
swagger: "2.0"
info:
  title: Social Media API
  description: API documentation
  version: 1.0.0
paths:
...
  /refresh:
    get:
      summary: Get new AccessToken
      operationId: RefreshToken
      tags: 
        - Users
      security:
        - cookieAuth: []
      responses:
        200:
          description: OK
...
  /forgetPassword:
    post:
      summary: Send reset email
      operationId: ForgetPassword
      tags:
        - Users
      parameters:
        - in: body
          name: email
          required: true
          schema:
            $ref: "#definitions/Email"
      responses:
        200:
          description: OK
  /resetPassword:
    put:
      summary: Reset forgotten Password
      description: Use the forgetPassword route above to send a token to your registered email address
      operationId: ResetPassword
      tags: [Users]
      parameters:
        - in: body
          name: passwordReset
          required: true
          schema:
            $ref: "#/definitions/PasswordReset"
      responses:
        200:
          description: OK
...
  /comments:
    get:
      summary: Get All Comments
      operationId: AllComments
      tags:
        - Comments
      security:
        - bearerAuth: []
      responses:
        200:
          description: OK

    post:
      summary: Comment on a Post
      operationId: CreateComment
      tags:
        - Comments
      parameters:
        - in: body
          name: Comment
          required: true
          schema:
            $ref: "#/definitions/Comment"
      security:
        - bearerAuth: []
      responses:
        200:
          description: OK
  /comments/{id}:
    get:
      summary: Get a Comment
      operationId: SingleComment
      tags:
        - Comments
      parameters:
        - $ref: "#/parameters/commentParams"
      security:
        - bearerAuth: []
      responses:
        200:
          description: OK

    delete:
      summary: Delete a Comment
      operationId: DeleteComment
      tags:
        - Comments
      parameters:
        - $ref: "#/parameters/commentParams"
      security:
        - bearerAuth: []
      responses:
        200:
          description: OK
  /comments/replies:
    post:
      summary: Reply a comment
      operationId: ReplyComment
      tags:
        - Comments
      parameters:
        - in: body
          name: Reply
          required: true
          schema:
            $ref: "#/definitions/Comment"
      security:
        - bearerAuth: []
      responses:
        200:
          description: OK
...
  /comments/totalReplies/{id}:
    get:
      summary: Get the total replies on a comment
      operationId: TotalReplies
      tags: [Comments]
      parameters:
        - $ref: "#/parameters/commentParams"
      security:
        - bearerAuth: []
      responses:
        200:
          description: OK
  /following:
    post:
      summary: Follow a User
      operationId: FollowUser
      tags:
        - Users
      parameters:
        - in: body
          name: User ID
          required: true
          schema:
            $ref: "#/definitions/Following"
      security:
        - bearerAuth: []
      responses:
        200:
          description: OK

    get:
      summary: View your following status
      operationId: Following
      tags: [Users]
      security:
        - bearerAuth: []
      responses:
        200:
          description: OK
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: jwt 
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
parameters:
...
  commentParams:
    in: path
    name: id
    required: true
    type: string
    minimum: 1
    description: The Comment's ID
...
definitions:
...
  Comment:
    type: object
    properties:
      postId:
        type: string
      text:
        type: string
    required:
      - postId
      - text
...
  Following:
    type: object
    properties:
      followingId:
        type: string
        example: The ID of the User you want to follow
    required:
      - followingId
...
  PasswordReset:
    type: object
    properties:
      password:
        type: string
        minLength: 8
      passwordConfirm:
        type: string
        minLength: 8
    required: [passwordCurrent, password, passwordConfirm]